Pytorch_BERT/requirements.txt
torchtext
attrdict
